application/requirements.txt
streamlit==1.28.0
psycopg2-binary==2.9.7
pandas==2.1.3
plotly==5.17.0
streamlit-option-menu==0.3.6